renamed component appCheckList to appCheckListModal
